pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["smoltrace"]

[project]
name = "smoltrace"
version = "0.0.3"
authors = [
    {name = "Kshitij Thakkar", email = "kshitijthakkar@rocketmail.com"}
]
description = "A comprehensive benchmarking and evaluation framework for smolagents."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
]
dependencies = [
    "smolagents>=1.0.0",
    "datasets",
    "huggingface_hub",
    "opentelemetry-sdk",
    "genai-otel-instrument[openinference]>=0.1.14",
    "duckduckgo-search",
    "ddgs"
]

[project.optional-dependencies]
gpu = ["genai-otel-instrument[gpu]>=0.1.14", "nvidia-ml-py"]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.21.1",

    # Code quality
    "black>=23.7.0",
    "isort>=5.12.0",
    "pylint>=2.17.5",
    "ruff>=0.0.285",
    "mypy>=1.4.1",
    "pre-commit>=3.5.0",

    # Build and release
    "build>=0.10.0",
    "twine>=4.0.2",

    # Security
    "safety>=2.3.5",
    "bandit>=1.7.5",
]

[project.urls]
"Homepage" = "https://github.com/Mandark-droid/SMOLTRACE"
"Bug Tracker" = "https://github.com/Mandark-droid/SMOLTRACE/issues"
"Documentation" = "https://github.com/Mandark-droid/SMOLTRACE#readme"
"Changelog" = "https://github.com/Mandark-droid/SMOLTRACE/blob/main/CHANGELOG.md"


[project.scripts]
smoltrace-eval = "smoltrace.cli:main"
smoltrace-cleanup = "smoltrace.cleanup:main"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
